update some code
allgemeine Sachen nur :)

+++ admin/lib/classes/user/login.php
@@ -131,7 +131,7 @@ class userLogin {
 	
 	public static function getUser() {
 		
-		return self::$userID;
+		return (int)self::$userID;
 		
 	}
